--- PetShop/src/packages/configs/Params.java
@@ -1,8 +1,0 @@
-package packages.configs;
-
-import java.net.InetAddress;
-public class Params {
-    public static final String HTTP_HOST = 'petstore.swagger.io';
-    public static final String HTTP_PROTOCOL = 'https';
-    public static final String GET_STATUS_PATH = 'v2/pet/findByStatus';
-}
