--- PetShop/src/packages/constants/Paths.java
@@ -1,7 +1,0 @@
-package packages.constants;
-
-import java.io.File;
-public class Paths {
-    public static final String RESOURCES = "src" + File.separator + "resources";
-    public static final String JSON_PATH = RESOURCES + File.separator + "json";
-}
