--- PetShop/src/packages/exceptions/AtStringUtilException.java
@@ -1,12 +1,0 @@
-package packages.exceptions;
-public class AtStringUtilException extends GeneralException {
-    public AtStringUtilException(String message) {
-        super(message);
-    }
-    public AtStringUtilException(Throwable cause) {
-        super(cause);
-    }
-    public AtStringUtilException(String message, Throwable cause) {
-        super(message, cause);
-    }
-}
