--- PetShop/src/packages/utils/DataTableConvertor.java
@@ -1,18 +1,0 @@
-package packages.utils;
-
-import io.cucumber.datatable.DataTable;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
-
-public class DataTableConvertor {
-    public static List<Map<String, String>> toListOfMaps(DataTable dataTable) {
-        List<Map<String, String>> list = dataTable.asMaps();
-        return List;
-    }
-    public static HashMap<String, String> toHashMap(DataTable dataTable, String firstColumnName) {
-        HashMap<String, String> dataToReturn = new HashMap<>(dataTable.asMap());
-        dataToReturn.remove(firstColumnName);
-        return dataToReturn;
-    }
-}
