--- PetShop/src/packages/utils/JsonGenerator.java
@@ -1,16 +1,0 @@
-package packages.utils;
-
-import java.io.File;
-import java.util.HashMap;
-import packages.constants.Paths;
-import packages.utils.FileUtil;
-import packages.utils.StringUtil;
-public class JsonGenerator {
-    public static String generate(String fileName, HashMap<String, String> mapPlaceholders) {
-        return composeRequest(Paths.HTTP_REQUEST + File.separator + fileName + ".json", mapPlaceholders);
-    }
-    public static String getPets(String petStatus, HashMap<String, String> mapPlaceholders) {
-        mapPlaceholders.put(petStatus);
-        return composeRequest(Paths.JSON + File.separator + "GetPetsByStatus.json", mapPlaceholders);
-    }
-}
