--- PetShop/src/resources/json/GetPetsByStatus.json
@@ -1,5 +1,0 @@
-{
-  "jsonData": {
-    "status": "#{placeholder_status}"
-  }
-}
