+++ src/test/java/configs/Params.java
@@ -1,0 +1,10 @@
+package configs;
+
+public class Params {
+
+    public static final String HTTP_HOST = "petstore.swagger.io";
+
+    public static final String HTTP_PROTOCOL = "https";
+
+    public static final String GET_STATUS_PATH = "v2/pet/findByStatus";
+}
