+++ src/test/java/constants/Paths.java
@@ -1,0 +1,10 @@
+package constants;
+
+import java.io.File;
+
+public class Paths {
+
+    public static final String RESOURCES = "src" + File.separator + "resources";
+
+    public static final String JSON_PATH = RESOURCES + File.separator + "json";
+}
