+++ src/test/java/exceptions/AtFileUtilException.java
@@ -1,0 +1,16 @@
+package exceptions;
+
+public class AtFileUtilException extends GeneralException {
+
+    public AtFileUtilException(String message) {
+        super(message);
+    }
+
+    public AtFileUtilException(Throwable cause) {
+        super(cause);
+    }
+
+    public AtFileUtilException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}
