+++ src/test/java/exceptions/AtStringUtilException.java
@@ -1,0 +1,16 @@
+package exceptions;
+
+public class AtStringUtilException extends GeneralException {
+
+    public AtStringUtilException(String message) {
+        super(message);
+    }
+
+    public AtStringUtilException(Throwable cause) {
+        super(cause);
+    }
+
+    public AtStringUtilException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}
