+++ src/test/java/exceptions/GeneralException.java
@@ -1,0 +1,16 @@
+package exceptions;
+
+public class GeneralException extends RuntimeException {
+
+    public GeneralException(String message) {
+        super(message);
+    }
+
+    public GeneralException(Throwable cause) {
+        super(cause);
+    }
+
+    public GeneralException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}
