+++ src/test/resources/json/GetPetById.json
@@ -1,0 +1,5 @@
+{
+  "jsonData": {
+    "id": "#{placeholder_status}"
+  }
+}
