+++ src/test/resources/json/GetPetsByStatus.json
@@ -1,0 +1,5 @@
+{
+  "jsonData": {
+    "status": ["#{placeholder_status}"]
+  }
+}
